pyproject.toml
[tool.poetry]
name = "learning-practice-django"
version = "1.0.0"
description = "Project for colleage learning practice on django for coffe shop \"Aogiri\""
authors = ["gh0sthck"]
license = "GNU"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.1"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
psycopg = "3.2"
celery = "^5.4.0"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
